Fusion PouvoirFigure et PouvoirFigureMini en un seul composant
- Ajout du variant "mini" dans PouvoirFigure (union discriminée TypeScript)
- Suppression de PouvoirFigureMini (remplacé par <PouvoirFigure variant="mini">)
- CollectiviteLocaleBlock mis à jour pour utiliser le nouveau variant
